# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Games API Doc",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080/Formation/GamesSite/BackEnd/public/api/",
            "description": "API Games DOC"
        }
    ],
    "paths": {
        "/games": {
            "get": {
                "tags": [
                    "Game"
                ],
                "summary": "Get a list of games",
                "operationId": "game",
                "responses": {
                    "404": {
                        "description": "No found games list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponseDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Return a list of Games",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GamesListDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/game/edit/{gameId}": {
            "put": {
                "tags": [
                    "Game"
                ],
                "summary": "Update Game",
                "operationId": "update",
                "parameters": [
                    {
                        "parameter": "gameId",
                        "name": "gameId",
                        "in": "path",
                        "description": "If for game to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update game",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/GameEditForm"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "No found Game, Platform, Editeur, Genre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponseDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Form is invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponseDTO"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponseDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Return a game detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameDetailDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/game/upload/{gameId}": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Upload image of game",
                "operationId": "upload",
                "parameters": [
                    {
                        "name": "gameId",
                        "in": "path",
                        "description": "Id of game",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Picture to upload",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Not found game or not found file to upload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponseDTO"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponseDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Return game upadted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameDetailDTO"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ApiErrorResponseDTO": {
                "title": "ApiErroResponseDTO",
                "description": "Model of error response",
                "properties": {
                    "statusCode": {
                        "description": "Status Error code",
                        "type": "integer"
                    },
                    "message": {
                        "description": "Error message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "GameDetailDTO": {
                "title": "GameDetailDTO",
                "description": "Model of game detail",
                "properties": {
                    "id": {
                        "title": "Game ID",
                        "description": "Id of the one game",
                        "type": "integer"
                    },
                    "nom": {
                        "title": "Game nom",
                        "description": "Nom of the game",
                        "type": "string"
                    },
                    "description": {
                        "title": "Game description",
                        "description": "Description of the Game",
                        "type": "string"
                    },
                    "prix": {
                        "title": "Game price",
                        "description": "Price of game",
                        "type": "decimal"
                    },
                    "image": {
                        "title": "Game picture",
                        "description": "Picture of game",
                        "type": "string"
                    },
                    "video": {
                        "title": "Game video",
                        "description": "Video of game",
                        "type": "string"
                    },
                    "editeur": {
                        "title": "Game editeur",
                        "description": "Editeur of the game",
                        "type": "string"
                    },
                    "genre": {
                        "title": "genre",
                        "description": "Genre of the game",
                        "type": "string"
                    },
                    "platform": {
                        "title": "platform",
                        "description": "Platform of the game",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "GamesListDTO": {
                "title": "GameListDTO",
                "description": "Model of games list response",
                "properties": {
                    "id": {
                        "title": "Game ID",
                        "description": "Id of the one game",
                        "type": "integer"
                    },
                    "nom": {
                        "title": "Game name",
                        "description": "Name of the one game",
                        "type": "string"
                    },
                    "description": {
                        "title": "Game Description",
                        "description": "Description of the one game",
                        "type": "string"
                    },
                    "image": {
                        "title": "Game image",
                        "description": "Image of the one game",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "GameEditForm": {
                "title": "GameEditForm",
                "description": "Model to update game",
                "properties": {
                    "nom": {
                        "title": "Game nom",
                        "description": "Nom of game",
                        "type": "string"
                    },
                    "description": {
                        "title": "Game description",
                        "description": "Game of description",
                        "type": "string"
                    },
                    "video": {
                        "title": "Game of video",
                        "description": "Videos of the game",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "idEditeur": {
                        "title": "Game editeur",
                        "description": "Id of the editeur of the game",
                        "type": "integer"
                    },
                    "idGenre": {
                        "title": "Game genre",
                        "description": "Id of the genre of the game",
                        "type": "integer"
                    },
                    "idPlatform": {
                        "title": "Game platform",
                        "description": "Id of the platform of the game",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    }
}